Allow verifying an Open VPN username and password authentication attempt
Additional configuration options have been added to allow an Open VPN administrator to decide which Duo challenges are enabled for different Open VPN authentication types - currently one of certificate or username and password - with the appropriate challenge types then being requested based on the authentication type actually requested by the user during connection attempts.

Where username and password is being used, only push or phone options are allowed to be configured, as the user has no way of entering a fall-back code or notifying that they would like to receive backup codes, with the administrator only being allowed to configure one of these two options since there's no way of a user selecting a preferred option during their authentication attempt. The certificate based authentication mechanism retains the same possible options as before, although the administrator now has an option for which mechanisms are enabled from the allowed list. For security, no authentication approaches are initially turned on, and no challenge mechanisms are enabled so that an administrator must select the specific options they want to have supported. If they fail to enable any options or enable an invalid combination of options then the plugin will log an appropriate error and return an error code to Open VPN which will result in the server failing to start. Similarly, checks are performed on the user's inputs and errors logged if any inputs are not appropriate for the currently enabled authentication and challenge mechanisms, with the challenge attempt then being aborted before calling the Duo APIs.

For backwards compatibility with old installations, any configuration that's using positional parameter parsing rather than a configuration file has username and password authentication disabled for challenges, and has all the certificate authentication challenges enabled.